+++ swim/macros/(Un-)Stun.js
@@ -107,7 +107,7 @@ async function main() {
             callback: (html) => { return; },
           }
         },
-        default: "No."
+        default: "one"
       }).render(true)
     }
     else {